improve type safety
